backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
  "albumentations>=2.0.8",
  "bs4>=0.0.2",
  "fastapi>=0.115.12",
  "firebase-admin>=6.9.0",
  "google-generativeai>=0.8.5",
  "google-search-results>=2.4.2",
  "googlesearch-python>=1.3.0",
  "langchain>=0.3.25",
  "langchain-chroma>=0.2.3",
  "langchain-community>=0.3.25",
  "logging>=0.4.9.6",
  "numpy>=1.26.0,<2.0.0",
  "opencv-python-headless>=4.11.0.86",
  "pandas>=2.3.0",
  "pillow>=11.2.1",
  "psycopg2-binary>=2.9.10",
  "pydantic[email]>=2.11.7",
  "requests>=2.32.4",
  "segmentation-models-pytorch>=0.5.0",
  "tensorflow-hub>=0.16.1",
  "tensorflow-text>=2.19.0",
  "tensorflow>=2.19.0",
  "tf-keras>=2.17.0",
  "timm>=1.0.15",
  "torch>=2.5.1",
  "torchvision>=0.20.1",
  "uvicorn>=0.34.3",
  "python-multipart>=0.0.20",
  "joblib>=1.5.1",
  "scikit-learn>=1.7.0",
  "duckduckgo-search>=8.0.4",
  "psutil>=5.9.0"
]


[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu121", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cu121", marker = "sys_platform == 'linux'" },
]
